Multiscale refactoring: Rename MultiscaleWebStore
It's not actually web-specific, and we want OME-zarr to work from web and file system
